Add logging and extend tests for StabilityAi services
Enhanced the logging features within the StabilityAiConfig by enabling a LoggingInterceptor. Introduced new tests for image resizing functionality and improved overall unit tests for StabilityAi services. Additionally, made some modifications to improve the quality and readability of the log outputs.

+++ src/main/java/com/kousenit/stabilityai/services/StabilityAiInterface.java
@@ -28,7 +28,7 @@ public interface StabilityAiInterface {
             contentType = MediaType.MULTIPART_FORM_DATA_VALUE)
     VideoId requestImageToVideo(@RequestPart Resource image,
                                 @RequestPart double cfg_scale,
-                                @RequestPart int motion_bucket_id);
+                                @RequestPart long motion_bucket_id);
 
     @GetExchange(value = "/image-to-video/result/{id}",
             accept = MediaType.APPLICATION_JSON_VALUE)
